Refactor Health Checks
Health checks are no longer necessary for `docker-dev` to function, but
they are a huge help in determining when an app server is ready to serve
requests. Before evaluating the results of the health check, determine
whether health checks take place on the container at all. If not, the
app should be ready when the container is up. This should allow for apps
that don't have a health check configured to still work with
`docker-dev`. Documentation has also been written around health
checks including why they might be useful for some cases.